Reconfigure how calendar and week data exist
This decouples themes from the calendar YAML file, and creates a more
manageable weekly data file that can expand to include agendas,
deadlines, etc.

Under this scheme, once set for a term, the calendar file can be left
unchanged. To switch, say, the plans for Week Ten with Week Eleven would
require just swapping the file names on `10.yml` and `11.yml`.
Significantly, it would *not* require swapping dates around (as would be
the case if dates were stored in the individual week files).